+++ TypedQuery/Column/NByteColumn.cs
@@ -153,7 +153,7 @@ namespace Sql.Column {
             return base.ToString();
         }
         public override System.Data.DbType DbType {
-            get { return System.Data.DbType.Int32; }
+            get { return System.Data.DbType.Byte; }
         }
         public override object? GetDefaultType() {
             return null;
